hpcviewer.e4: using part descriptor to generate view
to create multiple calling context views, we need to use part
descriptor.
However, there is a missing API in Eclipse e4 to specify a view: there
is no way to find a specific view if we have multiple views. Eclipse e4
will return all views of the same part descriptor.

We will deal with this later. At the moment the most important is to
generate view dynamically.