Fix bug for reading "corrupt" database
It is possible that hpcrun database is corrupt and a line statement is a
child of the program root as follows:

<program root> -> main.c:0

In this case, the flat-place holder is null since <program root> is a
fake procedure. We need to avoid to force a fake procedure to be a flat
holder.
Btw, since it's a garbage-in, the result is garbage-out. There's no
correct way to interpret this cct into flat view (it's impossible to
interpret a line statement without a procedure).