hpcdata: Fix bug #36 (omp idle is omitted in flat view)
The root cause of the bug is when parsing xml file, we do not
differentiate between procedure root (omp idle, program root, ...) and
an inline macro. 
Both procedures share the same characteristics: they don't have line
scope and their scope stack is empty.
One main difference is the inline macro is an alien, while procedure
root is not.