hpcviewer: Enable generating thread view from datacentric view
Thread view is a special view to display cct and metrics of a specific
thread (or a set of threads).
So far thread view is designed to display only from cct root. To enable
thread view from datacentric view, it requires major changes to display
from specific root.

The rationale for generating thread view from datacentric view is that
it can be useful to see different memory load/store for each thread (or
a group of threads).