hpcviewer: add root scope type as the 2nd ID in thread view
Thread view for cct and datacentric is different (as they have different
tree and root scope). 
By using the root scope type, thread view will be able to differentiate
between the thread view for cct and the one for datacentric. Thus, it's
possible to see side-by-side between the view of thread 1 from cct and
from datacentric.