hpcviewer: Fix the size of Scope column for Linux platform
On Linux, the Scope column of bottom-up and Flat views can be huge, and
the metric columns are located far to the right. This may confuse users
thinking there is no metrics available.
(this bug not happens on Mac or Windows, not sure with other Linux
distros because it depends on the installed gtk)

The hack to fix is to remove fix size column width, and depends entirely
with the column layout.