Draft for automatic selecting the first row
Selecting the first row is easily implemented in traditional table, but
very tricky for lazy content provider since we cannot select a row when
the table is not filled yet.

For dynamic views (bottom-up and flat view) are trickier since the table
is only ready if the view is visible or activated.

This commit allows to select the first row for lazy content provider,
and dynamic views. It works in Eclipse environment, but doesn't work on
dynamic views in production environment. Apparently Eclipse's
development environment uses different libraries compared to production
environment.